+++ onapp2vhi/main.py
@@ -11,6 +11,7 @@ from onapp2vhi.utilities.logs.logger import setup_logger
 from onapp2vhi.inc.vinfra_wrapper import VinfraFlavor
 from onapp2vhi.utilities.config_cli import ConfigCli
 from onapp2vhi.ops.migrate import migrate_impl
+from onapp2vhi.ops.update import update_vz_tools as update_vz_tools_impl
 
 
 cfg = None
@@ -214,17 +215,25 @@ def create_service_user():
     expose_value=True,
     help="Enable VM CPU and RAM hot plug",
 )
+@click.option(
+    "--network",
+    required=False,
+    type=str,
+    help="target appliance network on VHI"
+)
+# TODO! convert vz_guest_tools_install & clout_init_install to actual value
 def migrate(
-    user="",
-    vm="",
-    vm_ssh_port=22,
-    project="",
-    vz_guest_tools_install="true",
-    cloud_init_install="true",
-    placement="",
-    storage_policy="",
-    flavor="",
-    hotplug=False
+    user: str = "",
+    vm: str = "",
+    vm_ssh_port: int = 22,
+    project: str = "",
+    vz_guest_tools_install: str = "true",
+    cloud_init_install: str = "true",
+    placement: str = "",
+    storage_policy: str = "",
+    flavor: str = "",
+    hotplug: bool = False,
+    network: str = ""
 ):
     migrate_impl(
         cfg,
@@ -237,5 +246,11 @@ def migrate(
         placement=placement,
         storage_policy=storage_policy,
         flavor=flavor,
-        cpu_hotplug=hotplug
+        cpu_hotplug=hotplug,
+        network=network
     )
+
+
+@run.command()
+def update_vz_tools():
+    update_vz_tools_impl(cfg)