+++ tests/onapp2vhi/inc/test_onapp_helpers.py
@@ -22,7 +22,7 @@ from onapp2vhi.inc.onapp_helpers import (
     GenerateXmlConfig,
 )
 from onapp2vhi.inc.vinfra_wrapper import (
-    VinfraServer,
+    VinfraServiceComputeServer,
 )
 from onapp2vhi.inc.rest_client import OnAppRequests, OnAppRequestsException
 from onapp2vhi.utilities.config import OnApp2VHIConfig
@@ -1506,62 +1506,15 @@ class CreateNewVhiVmTestCase(OnAppHelpersTestCase):
                                    'test-server', 'localdomain', 'default')
         self.assertEqual(result, expected)
 
-    def test_create_multi_ips_list_iface_failed(self):
-        mock_disks = [{ 'size': '5',},]
-        mock_nics = [{ 'ips': ['1.1.1.1', '2.2.2.2'],},]
-
-        def mock_ssh_execute(command:str, real_data=False):
-            if 'service compute server create' in command:
-                return (0, json.dumps({'id': '1111-2222-3333-444444444444'}))
-            elif 'for ((i=1;i<=100;i++))' in command:
-                return (0, 'done')
-            elif 'service compute server iface list' in command:
-                return (1, 'list iface failed!')
-
-            raise RuntimeError(f'Unhandled command: {command}')
-
-        self.mock_ssh.execute.side_effect = mock_ssh_execute
-
-        result = create_new_vhi_vm(self.mock_cfg, self.mock_ssh, 'vinfra vhi-creds', 'abcdef',
-                                   'public2', 'vhi_image.qow', mock_disks, 'flavor_1_128', mock_nics,
-                                   'test-server', 'localdomain', 'default')
-        self.assertFalse(result)
-
-    def test_create_multi_ips_set_iface_failed(self):
-        mock_disks = [{ 'size': '5',},]
-        mock_nics = [{ 'ips': ['1.1.1.1', '2.2.2.2'],},]
-
-        def mock_ssh_execute(command:str, real_data=False):
-            if 'service compute server create' in command:
-                return (0, json.dumps({'id': '1111-2222-3333-444444444444'}))
-            elif 'for ((i=1;i<=100;i++))' in command:
-                return (0, 'done')
-            elif 'service compute server iface list' in command:
-                return (
-                    0,
-                    json.dumps([ { 'id': 'eth0', }, { 'id': 'eth1', } ])
-                )
-            elif 'service compute server iface set' in command:
-                return (1, 'not ok')
-
-            raise RuntimeError(f'Unhandled command: {command}')
-
-        self.mock_ssh.execute.side_effect = mock_ssh_execute
-
-        result = create_new_vhi_vm(self.mock_cfg, self.mock_ssh, 'vinfra vhi-creds', 'abcdef',
-                                   'public2', 'vhi_image.qow', mock_disks, 'flavor_1_128', mock_nics,
-                                   'test-server', 'localdomain', 'default')
-        self.assertFalse(result)
-
 
 class PrepareVhiMigrationDataTestCase(OnAppHelpersTestCase):
 
     def setUp(self):
         super().setUp()
         self.mock_onapprequests = Mock(spec=OnAppRequests)
-        self.mock_vinfraserver = Mock(spec=VinfraServer)
+        self.mock_vinfraserver = Mock(spec=VinfraServiceComputeServer)
 
-    @patch('onapp2vhi.inc.onapp_helpers.VinfraServer')
+    @patch('onapp2vhi.inc.onapp_helpers.VinfraServiceComputeServer')
     @patch('onapp2vhi.inc.onapp_helpers.OnAppRequests')
     def test_prepare_migration_data_all_users(self, mock_onapprequests, mock_vinfraserver):
         mock_onapprequests.return_value = self.mock_onapprequests
@@ -1659,7 +1612,7 @@ class PrepareVhiMigrationDataTestCase(OnAppHelpersTestCase):
         self.assertTrue(isinstance(result, list))
         self.assertEqual(result, expected)
 
-    @patch('onapp2vhi.inc.onapp_helpers.VinfraServer')
+    @patch('onapp2vhi.inc.onapp_helpers.VinfraServiceComputeServer')
     @patch('onapp2vhi.inc.onapp_helpers.OnAppRequests')
     def test_prepare_migration_data_all_users_no_vm(self, mock_onapprequests, mock_vinfraserver):
         mock_onapprequests.return_value = self.mock_onapprequests
@@ -1779,7 +1732,7 @@ class PrepareVhiMigrationDataTestCase(OnAppHelpersTestCase):
         self.assertTrue(isinstance(result, list))
         self.assertEqual(result, expected)
 
-    @patch('onapp2vhi.inc.onapp_helpers.VinfraServer')
+    @patch('onapp2vhi.inc.onapp_helpers.VinfraServiceComputeServer')
     @patch('onapp2vhi.inc.onapp_helpers.OnAppRequests')
     def test_prepare_migration_data_for_a_user(self, mock_onapprequests, mock_vinfraserver):
         mock_onapprequests.return_value = self.mock_onapprequests
@@ -1882,7 +1835,7 @@ class PrepareVhiMigrationDataTestCase(OnAppHelpersTestCase):
         self.assertTrue(isinstance(result, list))
         self.assertEqual(result, expected)
 
-    @patch('onapp2vhi.inc.onapp_helpers.VinfraServer')
+    @patch('onapp2vhi.inc.onapp_helpers.VinfraServiceComputeServer')
     @patch('onapp2vhi.inc.onapp_helpers.OnAppRequests')
     def test_prepare_migration_data_user_not_found(self, mock_onapprequests, mock_vinfraserver):
         mock_onapprequests.return_value = self.mock_onapprequests
@@ -1931,7 +1884,7 @@ class PrepareVhiMigrationDataTestCase(OnAppHelpersTestCase):
         result = prepare_vhi_migration_data(self.mock_cfg, 1)
         self.assertFalse(result)
 
-    @patch('onapp2vhi.inc.onapp_helpers.VinfraServer')
+    @patch('onapp2vhi.inc.onapp_helpers.VinfraServiceComputeServer')
     @patch('onapp2vhi.inc.onapp_helpers.OnAppRequests')
     def test_prepare_migration_data_user_has_no_vm(self, mock_onapprequests, mock_vinfraserver):
         mock_onapprequests.return_value = self.mock_onapprequests
@@ -1964,7 +1917,7 @@ class PrepareVhiMigrationDataTestCase(OnAppHelpersTestCase):
         result = prepare_vhi_migration_data(self.mock_cfg, 1)
         self.assertFalse(result)
 
-    @patch('onapp2vhi.inc.onapp_helpers.VinfraServer')
+    @patch('onapp2vhi.inc.onapp_helpers.VinfraServiceComputeServer')
     @patch('onapp2vhi.inc.onapp_helpers.OnAppRequests')
     def test_prepare_migration_data_with_vm_id(self, mock_onapprequests, mock_vinfraserver):
         mock_onapprequests.return_value = self.mock_onapprequests
@@ -2062,7 +2015,7 @@ class PrepareVhiMigrationDataTestCase(OnAppHelpersTestCase):
         result = prepare_vhi_migration_data(self.mock_cfg, vm_idn='abcdef')
         self.assertEqual(result, expected)
 
-    @patch('onapp2vhi.inc.onapp_helpers.VinfraServer')
+    @patch('onapp2vhi.inc.onapp_helpers.VinfraServiceComputeServer')
     @patch('onapp2vhi.inc.onapp_helpers.OnAppRequests')
     def test_prepare_migration_data_with_vm_not_found(self, mock_onapprequests, mock_vinfraserver):
         mock_onapprequests.return_value = self.mock_onapprequests
@@ -2084,7 +2037,7 @@ class PrepareVhiMigrationDataTestCase(OnAppHelpersTestCase):
         result = prepare_vhi_migration_data(self.mock_cfg, vm_idn='abcdef')
         self.assertFalse(result)
 
-    @patch("onapp2vhi.inc.onapp_helpers.VinfraServer")
+    @patch("onapp2vhi.inc.onapp_helpers.VinfraServiceComputeServer")
     @patch("onapp2vhi.inc.onapp_helpers.OnAppRequests")
     def test_prepare_migration_data_with_multiple_vm_id(
         self, mock_onapprequests, mock_vinfraserver